+++ client/src/api/GameAPI.ts
@@ -22,9 +22,9 @@ import { PixiManager } from './PixiManager';
 import { GAME_HEIGHT, GAME_WIDTH } from '../app/Pixi/GameBoard/GameBoard';
 import autoBind from 'auto-bind';
 import { findPath, getObstacles } from '../utils/Utils';
-import { Ship } from '../app/Pixi/Ship';
+import { Ship } from '../app/Pixi/Ships/Ship';
 import { PieceObject } from '../app/Pixi/PieceObject';
-import { Submarine } from '../app/Pixi/Submarine';
+import { Submarine } from '../app/Pixi/Ships/Submarine';
 
 export class GameAPI {
   private pixiManager: PixiManager;
