+++ client/src/app/Pixi/Ships/ShipManager.ts
@@ -1,7 +1,7 @@
-import { GameZIndex, PixiManager } from '../../api/PixiManager';
-import { PixiObject } from './PixiObject';
-import { BoardCoords } from './@PixiTypes';
-import { compareBoardCoords } from './PixiUtils';
+import { PixiManager, GameZIndex } from '../../../api/PixiManager';
+import { BoardCoords } from '../@PixiTypes';
+import { PixiObject } from '../PixiObject';
+import { compareBoardCoords } from '../PixiUtils';
 import { Ship } from './Ship';
 import { Submarine } from './Submarine';
